+++ lib/src/entities/events/guild_members_chunk_event.dart
@@ -1,8 +1,7 @@
 library entities;
 
-import 'package:discord/src/internal.dart';
-
 import '../../../entities.dart';
+import '../../internal.dart';
 
 class GuildMembersChunkEvent {
   final String guildId;