+++ lib/src/internal/discord_http_client.dart
@@ -1,10 +1,14 @@
 import 'dart:convert';
 
 import 'package:http/http.dart';
+import 'package:logging/logging.dart';
 
 import '../internal.dart';
 
 class DiscordHTTPClient {
+
+  static final log = Logger('discord.dart');
+
   static final _version = '0.0.1';
 
   static final _name = 'discord.dart';
@@ -16,11 +20,9 @@ class DiscordHTTPClient {
 
   final String _token;
 
-  late Client _client;
+  final Client _client;
 
-  DiscordHTTPClient(this._token) {
-    _client = Client();
-  }
+  DiscordHTTPClient(this._token) : _client = Client();
 
   List<T> Function(Map<String, dynamic>) listMapper<T>(
     T Function(Map<String, dynamic>) mapper,
@@ -38,7 +40,9 @@ class DiscordHTTPClient {
     String method = 'get',
   }) async {
     var uri = Uri.parse('$_baseUrl$endpoint');
+    log.finer(() => 'Requesting to $uri');
     if (query != null) {
+      log.finest(() => 'Using query parameters: $query');
       uri = uri.replace(queryParameters: query);
     }
     var headers = {
@@ -49,25 +53,31 @@ class DiscordHTTPClient {
     if (body != null) {
       headers['Content-Type'] = 'application/json';
       var stringBody = json.encode(body);
+      log.finest(() => 'Using body $stringBody');
       request.body = stringBody;
     }
     request.headers.addAll(headers);
 
+    log.fine('Sending request...');
     var response = await _client.send(request);
 
     if (response.statusCode == 204) {
+      log.fine('Discord responded with 204, returning null');
       return Future.value(null);
     }
 
-    // TODO handle all 2xx codes
     if (response.statusCode < 200 || response.statusCode >= 300) {
-      // TODO specialized exception
+      log.fine('Non 2xx response-code: ${response.statusCode}');
       var responseString = await response.stream.bytesToString();
+      log.finer(() => 'Error response string: $responseString');
+      // TODO specialized exception
       throw Exception(responseString);
     }
 
     var responseString = await response.stream.bytesToString();
+    log.fine('Loading API response as json data');
     var jsonData = json.decode(responseString);
+    log.fine('Calling converter and returning API result');
     return converter(jsonData is List ? {'arr': jsonData} : jsonData);
   }