+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 1.0.0
+## 0.0.2-beta1
 
-- Initial version, created by Stagehand
+- Add logging
+- Retrieve websocket uri from Discord
+- Code cleanup
+- Changed package name to lib_discord
+
+## 0.0.1-beta1
+
+First release